% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LS_script.R
\name{phase_plot}
\alias{phase_plot}
\title{Phase plot}
\usage{
phase_plot(t, x, period = 5)
}
\arguments{
\item{t}{independent}

\item{x}{input}

\item{period}{single value}
}
\value{
None, output is plot
}
\description{
This is the description
}
\examples{
t <- seq(from = 1, to = 10, length.out= 100)
x <- 2*sin(2*pi*t/5+0.5)+ 1 + rnorm(length(t),0, 0.25)
phase_plot(t,x, period=5)

}
